+++ server/utils/log-config.js
@@ -12,21 +12,23 @@
 
 const fs = require('fs');
 const path = require('path');
+const rotatingFileStream = require('rotating-file-stream');
+
+// CUstom modules
+const config = require('../config');
 
 const nodeEnv = process.env.NODE_ENV || 'development';
 // in the case of NODE_ENV=production, force logging to console.
 const nodeDebugLog = process.env.NODE_DEBUG_LOG || 0;
 
 let logToFile = (nodeEnv === 'production');
-let errorFilter = (nodeEnv === 'production');
+let errorFilter = ((nodeEnv === 'production') && (config.server.logFilter === 'error'));
 
 // enable console logging in production by export NODE_DEBUG_LOG=1
 if (nodeDebugLog) {
   logToFile = false;
   errorFilter = false;
 }
-// disable
-errorFilter = false;
 
 const logFolder = path.join(__dirname, '../../logs');
 const logFilename = logFolder + '/access.log';
@@ -66,12 +68,33 @@ if (logToFile) {
   // for start up message
   logStream = logFilename;
   // Function to write log entries to file as option property
-  logConfig.options.stream = fs.createWriteStream(logFilename, {
-    encoding: 'utf8',
-    mode: 0o644,
-    flags: 'a'
-  });
-}; // nodeEnv === production
+  if (((config.server.logRotateInterval) && (config.server.logRotateInterval.length > 1)) ||
+    ((config.server.logRotateSize) && (config.server.logRotateSize.length > 1))) {
+    const rotateOptions = {
+      encoding: 'utf8',
+      mode: 0o600,
+      rotate: 5
+    };
+    logStream += ' (Rotate';
+    if ((config.server.logRotateInterval) && (config.server.logRotateInterval.length > 1)) {
+      rotateOptions.interval = config.server.logRotateInterval;
+      logStream += ' interval:' + config.server.logRotateInterval;
+    }
+    if ((config.server.logRotateSize) && (config.server.logRotateSize.length > 1)) {
+      rotateOptions.size = config.server.logRotateSize;
+      logStream += ' size:' + config.server.logRotateSize;
+    }
+    logStream += ')';
+    logConfig.options.stream = rotatingFileStream.createStream(logFilename, rotateOptions);
+  } else {
+    // Function to write log entries to file as option property
+    logConfig.options.stream = fs.createWriteStream(logFilename, {
+      encoding: 'utf8',
+      mode: 0o600,
+      flags: 'a'
+    });
+  }
+};
 
 //
 // Filter function: If enabled, log only requrests with error status codes